@@ -74,7 +74,7 @@ run-coverage = "pytest test"
 run = "run-coverage --no-cov"
 
 [[tool.hatch.envs.test.matrix]]
-python = ["3.8", "3.9", "3.10", "3.11"]
+python = [ "3.10", "3.11", "3.12"]
 
 [tool.hatch.version]
 path = "./version.txt"
