@@ -15,7 +15,7 @@
 # specific language governing permissions and limitations
 # under the License.
 [build-system]
-requires = ["flit_core >=3.2,<4"]
+requires = ["flit_core >=3.11,<4"]
 build-backend = "flit_core.buildapi"
 
 [project]
@@ -41,45 +41,43 @@ classifiers = [
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
 ]
-requires-python = "~=3.9"
+requires-python = ">=3.9"
 
 dependencies = [
-    "black>=23.11.0",
-    # TODO: Click 8.2.0 is Python 3.10+ only by mistake (should be ~3.8) so we should limit it
-    #       until this is fixed https://github.com/pallets/click/issues/2896
-    "click>=8.1.8, <8.2.0",
+    "black>=25.0.0",
+    # There is a bug in Click 8.2.0 and 8.2.1 that makes flag variables do not properly evaluate
+    # "false"-y values set in environment variables. See https://github.com/pallets/click/issues/2952
+    "click>=8.1.8,<8.2.0",
     "filelock>=3.13.0",
-    #
-    # We pin flit in order to make sure reproducibility of provider packages is maintained
-    # It turns out that when packages are prepared metadata version in the produced packages
-    # is taken from the front-end not from the backend, so in order to make sure that the
-    # packages are reproducible, we should pin both backend in "build-system" and frontend in
-    # "dependencies" of the environment that is used to build the packages
-    "flit==3.12.0",
-    "flit-core==3.12.0",
+    "flit>=3.12.0",
+    "flit-core>=3.12.0",
     "google-api-python-client>=2.142.0",
     "google-auth-httplib2>=0.2.0",
     "google-auth-oauthlib>=1.2.0",
     "gitpython>=3.1.40",
-    "hatch==1.14.1",
+    "hatch>=1.14.1",
     "inputimeout>=1.0.4",
-    "jinja2>=3.1.0",
+    "jinja2>=3.1.5",
     "jsonschema>=4.19.1",
-    "packaging>=23.2",
-    "pre-commit>=3.5.0",
+    "packaging>=25.0",
+    "prek>=0.2.8",
     "psutil>=5.9.6",
     "pygithub>=2.1.1",
     "pytest-xdist>=3.3.1",
-    "pytest>=8.2,<9",
-    "pyyaml>=6.0.1",
-    "requests>=2.31.0",
-    "rich-click>=1.7.1",
+    "pytest>=8.3.3",
+    "pyyaml>=6.0.3",
+    "requests>=2.32.0",
+    "restructuredtext-lint>=1.4.0",
+    # Temporarily pinned to fix for changes in https://github.com/ewels/rich-click/releases/tag/v1.9.0
+    # Example failure without it: https://github.com/apache/airflow/actions/runs/17770084165/job/50505281673?pr=55725
+    "rich-click>=1.7.1,<1.9.0",
     "rich>=13.6.0",
-    "semver>=3.0.2",
     "tabulate>=0.9.0",
     "tomli>=2.0.1; python_version < '3.11'",
     "twine>=4.0.2",
-    "tqdm>=4.67.1"
+    "tqdm>=4.67.1",
+    "boto3>=1.34.90",
+    "semver>=3.0.4"
 ]
 
 [project.scripts]
@@ -90,7 +88,7 @@ name = "airflow_breeze"
 
 [tool.black]
 line-length = 110
-target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
+target-version = ['py39', 'py310', 'py311', 'py312']
 
 [tool.pytest.ini_options]
 addopts = "-rasl --verbosity=2 -p no:flaky -p no:nose -p no:legacypath"
