@@ -26,15 +26,15 @@ from sqlalchemy import asc, desc, func, select
 
 from airflow.api_connexion.exceptions import AlreadyExists, BadRequest, NotFound
 from airflow.api_connexion.parameters import check_limit, format_parameters
-from airflow.api_connexion.schemas.role_and_permission_schema import (
+from airflow.api_connexion.security import requires_access_custom_view
+from airflow.providers.fab.auth_manager.models import Action, Role
+from airflow.providers.fab.auth_manager.schemas.role_and_permission_schema import (
     ActionCollection,
     RoleCollection,
     action_collection_schema,
     role_collection_schema,
     role_schema,
 )
-from airflow.api_connexion.security import requires_access_custom_view
-from airflow.providers.fab.auth_manager.models import Action, Role
 from airflow.providers.fab.auth_manager.security_manager.override import FabAirflowSecurityManagerOverride
 from airflow.security import permissions
 from airflow.www.extensions.init_auth_manager import get_auth_manager
