@@ -23,6 +23,7 @@ import datetime
 # Copyright 2013, Daniel Vaz Gaspar
 from typing import TYPE_CHECKING
 
+import packaging.version
 from flask import current_app, g
 from flask_appbuilder.models.sqla import Model
 from sqlalchemy import (
@@ -32,6 +33,7 @@ from sqlalchemy import (
     ForeignKey,
     Index,
     Integer,
+    MetaData,
     String,
     Table,
     UniqueConstraint,
@@ -39,22 +41,33 @@ from sqlalchemy import (
     func,
     select,
 )
-from sqlalchemy.orm import backref, declared_attr, relationship
+from sqlalchemy.orm import backref, declared_attr, registry, relationship
 
+from airflow import __version__ as airflow_version
 from airflow.auth.managers.models.base_user import BaseUser
-from airflow.models.base import Base
-
-"""
-Compatibility note: The models in this file are duplicated from Flask AppBuilder.
-"""
-# Use airflow metadata to create the tables
-Model.metadata = Base.metadata
+from airflow.models.base import _get_schema, naming_convention
 
 if TYPE_CHECKING:
     try:
         from sqlalchemy import Identity
     except Exception:
         Identity = None
+
+"""
+Compatibility note: The models in this file are duplicated from Flask AppBuilder.
+"""
+
+metadata = MetaData(schema=_get_schema(), naming_convention=naming_convention)
+mapper_registry = registry(metadata=metadata)
+
+if packaging.version.parse(packaging.version.parse(airflow_version).base_version) >= packaging.version.parse(
+    "3.0.0"
+):
+    Model.metadata = metadata
+else:
+    from airflow.models.base import Base
+
+    Model.metadata = Base.metadata
 
 
 class Action(Model):
