@@ -36,9 +36,9 @@ AIRFLOW_SOURCES = Path(__file__).parents[3].resolve()
     [
         ("backend", "mysql", (True, ["sqlite", "mysql", "postgres", "none"]), None),
         ("backend", "xxx", (False, ["sqlite", "mysql", "postgres", "none"]), None),
-        ("python_major_minor_version", "3.9", (True, ["3.9", "3.10", "3.11", "3.12"]), None),
-        ("python_major_minor_version", "3.8", (False, ["3.9", "3.10", "3.11", "3.12"]), None),
-        ("python_major_minor_version", "3.7", (False, ["3.9", "3.10", "3.11", "3.12"]), None),
+        ("python_major_minor_version", "3.9", (False, ["3.10", "3.11", "3.12"]), None),
+        ("python_major_minor_version", "3.8", (False, ["3.10", "3.11", "3.12"]), None),
+        ("python_major_minor_version", "3.7", (False, ["3.10", "3.11", "3.12"]), None),
         ("missing", "value", None, AttributeError),
     ],
 )
