@@ -155,10 +155,12 @@ def test_find_matching_long_package_name_bad_filter():
             "fab",
             "",
             """
+    "apache-airflow-providers-common-compat>=1.2.1",
     "apache-airflow>=2.9.0",
-    "flask-appbuilder==4.5.2",
-    "flask-login>=0.6.2",
-    "flask>=2.2,<2.3",
+    "flask-appbuilder==4.5.3",
+    "flask-login>=0.6.3",
+    "flask-session>=0.8.0",
+    "flask>=2.2,<3",
     "google-re2>=1.0",
     "jmespath>=0.7.0",
     """,
@@ -168,10 +170,12 @@ def test_find_matching_long_package_name_bad_filter():
             "fab",
             "dev0",
             """
+    "apache-airflow-providers-common-compat>=1.2.1.dev0",
     "apache-airflow>=2.9.0.dev0",
-    "flask-appbuilder==4.5.2",
-    "flask-login>=0.6.2",
-    "flask>=2.2,<2.3",
+    "flask-appbuilder==4.5.3",
+    "flask-login>=0.6.3",
+    "flask-session>=0.8.0",
+    "flask>=2.2,<3",
     "google-re2>=1.0",
     "jmespath>=0.7.0",
     """,
@@ -181,10 +185,12 @@ def test_find_matching_long_package_name_bad_filter():
             "fab",
             "beta0",
             """
+    "apache-airflow-providers-common-compat>=1.2.1b0",
     "apache-airflow>=2.9.0b0",
-    "flask-appbuilder==4.5.2",
-    "flask-login>=0.6.2",
-    "flask>=2.2,<2.3",
+    "flask-appbuilder==4.5.3",
+    "flask-login>=0.6.3",
+    "flask-session>=0.8.0",
+    "flask>=2.2,<3",
     "google-re2>=1.0",
     "jmespath>=0.7.0",
     """,
@@ -486,7 +492,7 @@ def test_provider_jinja_context():
         "VERSION_SUFFIX": ".rc1",
         "PROVIDER_DESCRIPTION": "Amazon integration (including `Amazon Web Services (AWS) <https://aws.amazon.com/>`__).\n",
         "CHANGELOG_RELATIVE_PATH": "../../airflow/providers/amazon",
-        "SUPPORTED_PYTHON_VERSIONS": ["3.9", "3.10", "3.11", "3.12"],
+        "SUPPORTED_PYTHON_VERSIONS": ["3.10", "3.11", "3.12"],
         "PLUGINS": [],
         "MIN_AIRFLOW_VERSION": "2.9.0",
         "PROVIDER_REMOVED": False,
