@@ -113,10 +113,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("INTHEWILD.md",),
                 {
                     "selected-providers-list-as-string": None,
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "ci-image-build": "false",
                     "needs-helm-tests": "false",
                     "run-tests": "false",
@@ -147,10 +147,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
             pytest.param(
                 ("tests/api/file.py",),
                 {
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -174,10 +174,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("airflow/operators/file.py",),
                 {
                     "selected-providers-list-as-string": None,
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -201,10 +201,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("airflow/serialization/python.py",),
                 {
                     "selected-providers-list-as-string": None,
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -228,10 +228,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("docs/file.rst",),
                 {
                     "selected-providers-list-as-string": None,
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -259,10 +259,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ),
                 {
                     "selected-providers-list-as-string": None,
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "needs-helm-tests": "true",
@@ -287,10 +287,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("generated/provider_dependencies.json",),
                 {
                     "selected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
-                    "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
-                    "python-versions": "['3.9', '3.10', '3.11', '3.12']",
-                    "python-versions-list-as-string": "3.9 3.10 3.11 3.12",
+                    "all-python-versions": "['3.10', '3.11', '3.12']",
+                    "all-python-versions-list-as-string": "3.10 3.11 3.12",
+                    "python-versions": "['3.10', '3.11', '3.12']",
+                    "python-versions-list-as-string": "3.10 3.11 3.12",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "needs-helm-tests": "true",
@@ -314,10 +314,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("generated/provider_dependencies.json",),
                 {
                     "selected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
-                    "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
-                    "python-versions": "['3.9', '3.10', '3.11', '3.12']",
-                    "python-versions-list-as-string": "3.9 3.10 3.11 3.12",
+                    "all-python-versions": "['3.10', '3.11', '3.12']",
+                    "all-python-versions-list-as-string": "3.10 3.11 3.12",
+                    "python-versions": "['3.10', '3.11', '3.12']",
+                    "python-versions-list-as-string": "3.10 3.11 3.12",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "needs-helm-tests": "true",
@@ -340,10 +340,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("tests/utils/test_cli_util.py",),
                 {
                     "selected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "needs-helm-tests": "true",
@@ -366,10 +366,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("tests_common/__init__.py",),
                 {
                     "selected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "needs-helm-tests": "true",
@@ -394,10 +394,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("airflow/ui/src/index.tsx",),
                 {
                     "selected-providers-list-as-string": None,
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "ci-image-build": "false",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -560,15 +560,15 @@ def test_full_test_needed_when_scripts_changes(files: tuple[str, ...], expected_
                 "main",
                 {
                     "all-versions": "true",
-                    "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
-                    "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
+                    "all-python-versions": "['3.10', '3.11', '3.12']",
+                    "all-python-versions-list-as-string": "3.10 3.11 3.12",
                     "mysql-versions": "['8.0', '8.4']",
                     "postgres-versions": "['13', '14', '15', '16', '17']",
-                    "python-versions": "['3.9', '3.10', '3.11', '3.12']",
-                    "python-versions-list-as-string": "3.9 3.10 3.11 3.12",
+                    "python-versions": "['3.10', '3.11', '3.12']",
+                    "python-versions-list-as-string": "3.10 3.11 3.12",
                     "kubernetes-versions": "['v1.28.15', 'v1.29.12', 'v1.30.8', 'v1.31.4', 'v1.32.0']",
                     "kubernetes-versions-list-as-string": "v1.28.15 v1.29.12 v1.30.8 v1.31.4 v1.32.0",
-                    "kubernetes-combos-list-as-string": "3.9-v1.28.15 3.10-v1.29.12 3.11-v1.30.8 3.12-v1.31.4 3.9-v1.32.0",
+                    "kubernetes-combos-list-as-string": "3.10-v1.28.15 3.11-v1.29.12 3.12-v1.30.8 3.10-v1.31.4 3.11-v1.32.0",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "run-tests": "true",
@@ -595,16 +595,16 @@ def test_full_test_needed_when_scripts_changes(files: tuple[str, ...], expected_
                 "main",
                 {
                     "selected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
                     "all-versions": "false",
                     "mysql-versions": "['8.0']",
                     "postgres-versions": "['13']",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "kubernetes-versions": "['v1.28.15']",
                     "kubernetes-versions-list-as-string": "v1.28.15",
-                    "kubernetes-combos-list-as-string": "3.9-v1.28.15",
+                    "kubernetes-combos-list-as-string": "3.10-v1.28.15",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "run-tests": "true",
@@ -631,16 +631,16 @@ def test_full_test_needed_when_scripts_changes(files: tuple[str, ...], expected_
                 "main",
                 {
                     "selected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
                     "all-versions": "false",
                     "mysql-versions": "['8.0']",
                     "postgres-versions": "['13']",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "kubernetes-versions": "['v1.28.15']",
                     "kubernetes-versions-list-as-string": "v1.28.15",
-                    "kubernetes-combos-list-as-string": "3.9-v1.28.15",
+                    "kubernetes-combos-list-as-string": "3.10-v1.28.15",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "run-tests": "true",
@@ -707,14 +707,14 @@ def test_full_test_needed_when_scripts_changes(files: tuple[str, ...], expected_
                 "main",
                 {
                     "selected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
                     "all-versions": "false",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "kubernetes-versions": "['v1.28.15']",
                     "kubernetes-versions-list-as-string": "v1.28.15",
-                    "kubernetes-combos-list-as-string": "3.9-v1.28.15",
+                    "kubernetes-combos-list-as-string": "3.10-v1.28.15",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "run-tests": "true",
@@ -741,14 +741,14 @@ def test_full_test_needed_when_scripts_changes(files: tuple[str, ...], expected_
                 "main",
                 {
                     "selected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
                     "all-versions": "false",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "kubernetes-versions": "['v1.28.15']",
                     "kubernetes-versions-list-as-string": "v1.28.15",
-                    "kubernetes-combos-list-as-string": "3.9-v1.28.15",
+                    "kubernetes-combos-list-as-string": "3.10-v1.28.15",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "run-tests": "true",
@@ -775,10 +775,10 @@ def test_full_test_needed_when_scripts_changes(files: tuple[str, ...], expected_
                 ("full tests needed",),
                 "v2-7-stable",
                 {
-                    "all-python-versions": "['3.9']",
-                    "all-python-versions-list-as-string": "3.9",
-                    "python-versions": "['3.9']",
-                    "python-versions-list-as-string": "3.9",
+                    "all-python-versions": "['3.10']",
+                    "all-python-versions-list-as-string": "3.10",
+                    "python-versions": "['3.10']",
+                    "python-versions-list-as-string": "3.10",
                     "all-versions": "false",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
@@ -788,7 +788,7 @@ def test_full_test_needed_when_scripts_changes(files: tuple[str, ...], expected_
                     "docs-build": "true",
                     "docs-list-as-string": "apache-airflow docker-stack",
                     "full-tests-needed": "true",
-                    "skip-pre-commits": "check-airflow-provider-compatibility,check-extra-packages-references,check-provider-yaml-valid,identity,lint-helm-chart,mypy-airflow,mypy-dev,mypy-docs,mypy-providers,validate-operators-init",
+                    "skip-pre-commits": "check-airflow-provider-compatibility,check-extra-packages-references,check-provider-yaml-valid,identity,kubeconform,lint-helm-chart,mypy-airflow,mypy-dev,mypy-docs,mypy-providers,validate-operators-init",
                     "upgrade-to-newer-dependencies": "false",
                     "core-test-types-list-as-string": "API Always CLI Core Operators Other "
                     "Serialization WWW",
@@ -824,8 +824,8 @@ def test_expected_output_full_tests_needed(
             ("INTHEWILD.md",),
             {
                 "selected-providers-list-as-string": None,
-                "all-python-versions": "['3.9']",
-                "all-python-versions-list-as-string": "3.9",
+                "all-python-versions": "['3.10']",
+                "all-python-versions-list-as-string": "3.10",
                 "ci-image-build": "false",
                 "needs-helm-tests": "false",
                 "run-tests": "false",
@@ -847,8 +847,8 @@ def test_expected_output_full_tests_needed(
                 "tests/providers/google/file.py",
             ),
             {
-                "all-python-versions": "['3.9']",
-                "all-python-versions-list-as-string": "3.9",
+                "all-python-versions": "['3.10']",
+                "all-python-versions-list-as-string": "3.10",
                 "needs-helm-tests": "false",
                 "ci-image-build": "true",
                 "prod-image-build": "true",
@@ -874,8 +874,8 @@ def test_expected_output_full_tests_needed(
                 "tests/providers/google/file.py",
             ),
             {
-                "all-python-versions": "['3.9']",
-                "all-python-versions-list-as-string": "3.9",
+                "all-python-versions": "['3.10']",
+                "all-python-versions-list-as-string": "3.10",
                 "ci-image-build": "true",
                 "prod-image-build": "true",
                 "needs-helm-tests": "false",
@@ -899,8 +899,8 @@ def test_expected_output_full_tests_needed(
                 "tests/providers/google/file.py",
             ),
             {
-                "all-python-versions": "['3.9']",
-                "all-python-versions-list-as-string": "3.9",
+                "all-python-versions": "['3.10']",
+                "all-python-versions-list-as-string": "3.10",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -944,8 +944,8 @@ def test_expected_output_pull_request_v2_7(
             "main",
             {
                 "selected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
-                "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
+                "all-python-versions": "['3.10', '3.11', '3.12']",
+                "all-python-versions-list-as-string": "3.10 3.11 3.12",
                 "ci-image-build": "true",
                 "prod-image-build": "true",
                 "needs-helm-tests": "true",
@@ -965,14 +965,14 @@ def test_expected_output_pull_request_v2_7(
             (),
             "v2-3-stable",
             {
-                "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
-                "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
+                "all-python-versions": "['3.10', '3.11', '3.12']",
+                "all-python-versions-list-as-string": "3.10 3.11 3.12",
                 "ci-image-build": "true",
                 "prod-image-build": "true",
                 "needs-helm-tests": "false",
                 "run-tests": "true",
                 "docs-build": "true",
-                "skip-pre-commits": "check-airflow-provider-compatibility,check-extra-packages-references,check-provider-yaml-valid,identity,lint-helm-chart,mypy-airflow,mypy-dev,mypy-docs,mypy-providers,validate-operators-init",
+                "skip-pre-commits": "check-airflow-provider-compatibility,check-extra-packages-references,check-provider-yaml-valid,identity,kubeconform,lint-helm-chart,mypy-airflow,mypy-dev,mypy-docs,mypy-providers,validate-operators-init",
                 "docs-list-as-string": "apache-airflow docker-stack",
                 "upgrade-to-newer-dependencies": "true",
                 "core-test-types-list-as-string": "API Always CLI Core Operators Other Serialization WWW",
@@ -988,8 +988,8 @@ def test_expected_output_pull_request_v2_7(
             "main",
             {
                 "selected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
-                "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
+                "all-python-versions": "['3.10', '3.11', '3.12']",
+                "all-python-versions-list-as-string": "3.10 3.11 3.12",
                 "ci-image-build": "true",
                 "prod-image-build": "true",
                 "needs-helm-tests": "true",
@@ -1029,8 +1029,8 @@ def test_expected_output_push(
             ("INTHEWILD.md",),
             {
                 "selected-providers-list-as-string": None,
-                "all-python-versions": "['3.9']",
-                "all-python-versions-list-as-string": "3.9",
+                "all-python-versions": "['3.10']",
+                "all-python-versions-list-as-string": "3.10",
                 "ci-image-build": "false",
                 "needs-helm-tests": "false",
                 "run-tests": "false",
@@ -1051,8 +1051,8 @@ def test_expected_output_push(
             ("tests/system/any_file.py",),
             {
                 "selected-providers-list-as-string": None,
-                "all-python-versions": "['3.9']",
-                "all-python-versions-list-as-string": "3.9",
+                "all-python-versions": "['3.10']",
+                "all-python-versions-list-as-string": "3.10",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -1073,8 +1073,8 @@ def test_expected_output_push(
         pytest.param(
             ("airflow/models/test.py",),
             {
-                "all-python-versions": "['3.9']",
-                "all-python-versions-list-as-string": "3.9",
+                "all-python-versions": "['3.10']",
+                "all-python-versions-list-as-string": "3.10",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -1096,8 +1096,8 @@ def test_expected_output_push(
         pytest.param(
             ("airflow/file.py",),
             {
-                "all-python-versions": "['3.9']",
-                "all-python-versions-list-as-string": "3.9",
+                "all-python-versions": "['3.10']",
+                "all-python-versions-list-as-string": "3.10",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -1153,8 +1153,8 @@ def test_no_commit_provided_trigger_full_build_for_any_event_type(github_event):
     )
     assert_outputs_are_printed(
         {
-            "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
-            "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
+            "all-python-versions": "['3.10', '3.11', '3.12']",
+            "all-python-versions-list-as-string": "3.10 3.11 3.12",
             "ci-image-build": "true",
             "prod-image-build": "true",
             "needs-helm-tests": "true",
@@ -1189,8 +1189,8 @@ def test_files_provided_trigger_full_build_for_any_event_type(github_event):
     )
     assert_outputs_are_printed(
         {
-            "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
-            "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
+            "all-python-versions": "['3.10', '3.11', '3.12']",
+            "all-python-versions-list-as-string": "3.10 3.11 3.12",
             "ci-image-build": "true",
             "prod-image-build": "true",
             "needs-helm-tests": "true",
