@@ -43,7 +43,7 @@ PRE_INSTALLED_PROVIDERS = [
     "common.compat",
     "common.io",
     "common.sql",
-    "fab>=1.0.2",
+    "fab>=1.5.4,<3.0.0",
     "ftp",
     "http",
     "imap",
@@ -156,7 +156,8 @@ CORE_EXTRAS: dict[str, list[str]] = {
 
 DOC_EXTRAS: dict[str, list[str]] = {
     "doc": [
-        "astroid>=3; python_version >= '3.9'",
+        # Astroid 4 released 5 Oct 2025 breaks autoapi https://github.com/apache/airflow/issues/56420
+        "astroid>=3,<4; python_version >= '3.9'",
         "checksumdir>=1.2.0; python_version >= '3.9'",
         "click>=8.1.8; python_version >= '3.9'",
         "docutils>=0.21; python_version >= '3.9'",
@@ -201,7 +202,9 @@ DEVEL_EXTRAS: dict[str, list[str]] = {
         "pipdeptree>=2.13.1",
         "pygithub>=2.1.1",
         "restructuredtext-lint>=1.4.0",
-        "rich-click>=1.7.0",
+        # Temporarily pinned to fix for changes in https://github.com/ewels/rich-click/releases/tag/v1.9.0
+        # Example failure without it: https://github.com/apache/airflow/actions/runs/17770084165/job/50505281673?pr=55725
+        "rich-click>=1.7.1,<1.9.0",
         "semver>=3.0.2",
         "towncrier>=23.11.0",
         "twine>=4.0.2",
@@ -227,8 +230,8 @@ DEVEL_EXTRAS: dict[str, list[str]] = {
         "types-aiofiles",
         "types-certifi",
         "types-croniter",
-        "types-docutils",
-        "types-paramiko",
+        "types-docutils<0.22.0",
+        "types-paramiko<4.0.0",
         "types-protobuf",
         "types-python-dateutil",
         "types-python-slugify",
@@ -266,9 +269,10 @@ DEVEL_EXTRAS: dict[str, list[str]] = {
         "pytest-rerunfailures>=13.0",
         "pytest-timeouts>=1.2.1",
         "pytest-xdist>=3.5.0",
-        "pytest>=8.2,<9",
+        "pytest>=8.2,<8.3",
         "requests_mock>=1.11.0",
-        "time-machine>=2.13.0",
+        # Time machine is only used in tests and version 3 introduced breaking changes
+        "time-machine>=2.13.0,<3",
         "wheel>=0.42.0",
     ],
     "devel": [
@@ -406,8 +410,7 @@ DEPRECATED_EXTRAS: dict[str, list[str]] = {
 DEPENDENCIES = [
     # Alembic is important to handle our migrations in predictable and performant way. It is developed
     # together with SQLAlchemy. Our experience with Alembic is that it very stable in minor version
-    # The 1.13.0 of alembic marked some migration code as SQLAlchemy 2+ only so we limit it to 1.13.1
-    "alembic>=1.13.1, <2.0",
+    "alembic>=1.14.0, <2.0",
     "argcomplete>=1.10",
     "asgiref>=2.3.0",
     "attrs>=22.1.0",
@@ -416,13 +419,7 @@ DEPENDENCIES = [
     "blinker>=1.6.2",
     "colorlog>=6.8.2",
     "configupdater>=3.1.1",
-    # `airflow/www/extensions/init_views` imports `connexion.decorators.validation.RequestBodyValidator`
-    # connexion v3 has refactored the entire module to middleware, see: /spec-first/connexion/issues/1525
-    # Specifically, RequestBodyValidator was removed in: /spec-first/connexion/pull/1595
-    # The usage was added in #30596, seemingly only to override and improve the default error message.
-    # Either revert that change or find another way, preferably without using connexion internals.
-    # This limit can be removed after https://github.com/apache/airflow/issues/35234 is fixed
-    "connexion[flask]>=2.14.2,<3.0",
+    "connexion>=2.15.1,<3.0",
     "cron-descriptor>=1.2.24",
     "croniter>=2.0.2",
     "cryptography>=41.0.0",
@@ -432,15 +429,10 @@ DEPENDENCIES = [
     # description on PyPI for more details: https://pypi.org/project/eval-type-backport/
     # see https://github.com/pydantic/pydantic/issues/10958
     'eval-type-backport>=0.2.0;python_version<"3.10"',
-    "flask-caching>=2.0.0",
-    # Flask-Session 0.6 add new arguments into the SqlAlchemySessionInterface constructor as well as
-    # all parameters now are mandatory which make AirflowDatabaseSessionInterface incompatible with this version.
-    "flask-session>=0.4.0,<0.6",
-    "flask-wtf>=1.1.0",
-    # Flask 2.3 is scheduled to introduce a number of deprecation removals - some of them might be breaking
-    # for our dependencies - notably `_app_ctx_stack` and `_request_ctx_stack` removals.
-    # We should remove the limitation after 2.3 is released and our dependencies are updated to handle it
-    "flask>=2.2.1,<2.3",
+    "flask-caching>=2.3.1",
+    "flask-session>=0.8.0",
+    "flask-wtf>=1.2.2",
+    "flask>=2.3.3,<4",
     "fsspec>=2023.10.0",
     'google-re2>=1.0;python_version<"3.12"',
     'google-re2>=1.1;python_version>="3.12"',
@@ -491,12 +483,9 @@ DEPENDENCIES = [
     "tabulate>=0.7.5",
     "tenacity>=8.0.0,!=8.2.0",
     "termcolor>=1.1.0",
-    # Universal Pathlib 0.2.4 adds extra validation for Paths and our integration with local file paths
-    # Does not work with it Tracked in https://github.com/fsspec/universal_pathlib/issues/276
-    "universal-pathlib>=0.2.2,!=0.2.4",
-    # Werkzug 3 breaks Flask-Login 0.6.2, also connexion needs to be updated to >= 3.0
-    # we should remove this limitation when FAB supports Flask 2.3 and we migrate connexion to 3+
-    "werkzeug>=2.0,<3",
+    # https://github.com/apache/airflow/issues/56369 , rework universal-pathlib usage
+    "universal-pathlib>=0.2.6,<0.3.0",
+    "werkzeug>=3.1.3,<4",
 ]
 
 
@@ -586,6 +575,7 @@ def get_provider_requirement(provider_spec: str) -> str:
 
         current_airflow_version = Version(airflow_version_match.group(1))
         provider_id, min_version = provider_spec.split(">=")
+        min_version = min_version.split(",")[0]
         provider_version = Version(min_version)
         if provider_version.is_prerelease and not current_airflow_version.is_prerelease:
             # strip pre-release version from the pre-installed provider's version when we are preparing
