@@ -21,10 +21,10 @@ from unittest.mock import Mock, patch
 import pytest
 from flask import Response
 from flask_appbuilder.const import AUTH_LDAP
+from tests_common.test_utils.compat import AIRFLOW_V_2_9_PLUS
 
-from airflow.api.auth.backend.basic_auth import requires_authentication
+from airflow.providers.fab.auth_manager.api.auth.backend.basic_auth import requires_authentication
 from airflow.www import app as application
-from tests.test_utils.compat import AIRFLOW_V_2_9_PLUS
 
 pytestmark = [
     pytest.mark.skipif(not AIRFLOW_V_2_9_PLUS, reason="Tests for Airflow 2.9.0+ only"),
