@@ -16,10 +16,7 @@
 # under the License.
 from __future__ import annotations
 
-from airflow.api.auth.backend.kerberos_auth import (
-    init_app as base_init_app,
-)
-from tests.test_utils.compat import ignore_provider_compatibility_error
+from tests_common.test_utils.compat import ignore_provider_compatibility_error
 
 with ignore_provider_compatibility_error("2.9.0+", __file__):
     from airflow.providers.fab.auth_manager.api.auth.backend.kerberos_auth import init_app
@@ -27,4 +24,4 @@ with ignore_provider_compatibility_error("2.9.0+", __file__):
 
 class TestKerberosAuth:
     def test_init_app(self):
-        assert init_app == base_init_app
+        init_app
