@@ -20,9 +20,9 @@ from __future__ import annotations
 from unittest import mock
 
 import pytest
+from tests_common.test_utils.compat import ignore_provider_compatibility_error
 
 from airflow.cli import cli_parser
-from tests.test_utils.compat import ignore_provider_compatibility_error
 
 with ignore_provider_compatibility_error("2.9.0+", __file__):
     from airflow.providers.fab.auth_manager.cli_commands import sync_perm_command
