@@ -24,9 +24,9 @@ from contextlib import redirect_stdout
 from io import StringIO
 
 import pytest
+from tests_common.test_utils.compat import ignore_provider_compatibility_error
 
 from airflow.cli import cli_parser
-from tests.test_utils.compat import ignore_provider_compatibility_error
 
 with ignore_provider_compatibility_error("2.9.0+", __file__):
     from airflow.providers.fab.auth_manager.cli_commands import user_command
