@@ -17,10 +17,10 @@
 from __future__ import annotations
 
 import pytest
+from tests_common.test_utils.config import conf_vars
+from tests_common.test_utils.decorators import dont_initialize_flask_app_submodules
 
 from airflow.www import app
-from tests.test_utils.config import conf_vars
-from tests.test_utils.decorators import dont_initialize_flask_app_submodules
 
 
 @pytest.fixture(scope="session")
@@ -28,13 +28,27 @@ def minimal_app_for_auth_api():
     @dont_initialize_flask_app_submodules(
         skip_all_except=[
             "init_appbuilder",
-            "init_api_experimental_auth",
+            "init_api_auth",
             "init_api_auth_provider",
+            "init_api_connexion",
             "init_api_error_handlers",
+            "init_airflow_session_interface",
+            "init_appbuilder_views",
         ]
     )
     def factory():
-        with conf_vars({("api", "auth_backends"): "tests.test_utils.remote_user_api_auth_backend"}):
+        with conf_vars(
+            {
+                (
+                    "api",
+                    "auth_backends",
+                ): "providers.tests.fab.auth_manager.api_endpoints.remote_user_api_auth_backend,airflow.providers.fab.auth_manager.api.auth.backend.session",
+                (
+                    "core",
+                    "auth_manager",
+                ): "airflow.providers.fab.auth_manager.fab_auth_manager.FabAuthManager",
+            }
+        ):
             _app = app.create_app(testing=True, config={"WTF_CSRF_ENABLED": False})  # type:ignore
             _app.config["AUTH_ROLE_PUBLIC"] = None
             return _app
@@ -43,7 +57,7 @@ def minimal_app_for_auth_api():
 
 
 @pytest.fixture
-def set_auto_role_public(request):
+def set_auth_role_public(request):
     app = request.getfixturevalue("minimal_app_for_auth_api")
     auto_role_public = app.config["AUTH_ROLE_PUBLIC"]
     app.config["AUTH_ROLE_PUBLIC"] = request.param
@@ -51,3 +65,11 @@ def set_auto_role_public(request):
     yield
 
     app.config["AUTH_ROLE_PUBLIC"] = auto_role_public
+
+
+@pytest.fixture(scope="module")
+def dagbag():
+    from airflow.models import DagBag
+
+    DagBag(include_examples=True, read_dags_from_db=False).sync_to_db()
+    return DagBag(include_examples=True, read_dags_from_db=True)
