@@ -19,9 +19,9 @@ from __future__ import annotations
 from unittest.mock import Mock, patch
 
 import pytest
+from tests_common.test_utils.compat import ignore_provider_compatibility_error
 
 from airflow.security.permissions import ACTION_CAN_READ, RESOURCE_DAG
-from tests.test_utils.compat import ignore_provider_compatibility_error
 
 permissions = [(ACTION_CAN_READ, RESOURCE_DAG)]
 
