@@ -16,7 +16,7 @@
 # under the License.
 from __future__ import annotations
 
-from tests.test_utils.compat import ignore_provider_compatibility_error
+from tests_common.test_utils.compat import ignore_provider_compatibility_error
 
 with ignore_provider_compatibility_error("2.9.0+", __file__):
     from airflow.providers.fab.auth_manager.security_manager.constants import EXISTING_ROLES
