@@ -19,7 +19,7 @@ from __future__ import annotations
 from unittest import mock
 from unittest.mock import Mock
 
-from tests.test_utils.compat import ignore_provider_compatibility_error
+from tests_common.test_utils.compat import ignore_provider_compatibility_error
 
 with ignore_provider_compatibility_error("2.9.0+", __file__):
     from airflow.providers.fab.auth_manager.security_manager.override import FabAirflowSecurityManagerOverride
